become: run gvim in nofork mode
If running vim in one of its graphical modes as a subcommand, add the
--nofork command to prevent the editor from going into the background.
Otherwise, we end up destroying the temporary tickets and tokens
immediately and the file can't be edited.

This is a kludgy and incomplete solution. In particular, it means the
command doesn't terminate until the editor does. We probably should
put become into the background in this case, so the user gets his shell
back.