timelog-plus: avoid logging negative time
When there are already tasks on the stack at startup, time starts
accumulating for the topmost task at startup; the time since the program
was last running is treated as an implied break.  However, it is possible
to push a new task with a timestamp during the implied break (i.e.
before the program was restarted).

In this situation, no time should be logged for the task that was on top
at startup, since it turns out not to ever have been the topmost task.